+++ src/Spaces/Jacobi/jacobitransform.jl
@@ -8,11 +8,13 @@ immutable JacobiTransformPlan{DD,T,TT}
 end
 
 plan_transform(S::Jacobi,v::Vector) = JacobiTransformPlan(S,gaussjacobi(length(v),S.a,S.b)...)
-function transform(S::Jacobi,vals,plan::JacobiTransformPlan)
+function *(plan::JacobiTransformPlan,vals)
 #    @assert S==plan.space
+    S = plan.space
     x,w = plan.points, plan.weights
     V=jacobip(0:length(vals)-1,S.a,S.b,x)'
-    w2=w.*(1-x).^(S.a-plan.space.a).*(1+x).^(S.b-plan.space.b)   # need to weight if plan is different
+#    w2=w.*(1-x).^(S.a-plan.space.a).*(1+x).^(S.b-plan.space.b)   # need to weight if plan is different
+    w2=w
     nrm=(V.^2)*w2
 
     V*(w2.*vals)./nrm
@@ -28,10 +30,9 @@ end
 
 
 plan_itransform(S::Jacobi,cfs::Vector) = JacobiITransformPlan(S,points(S,length(cfs)))
-function itransform(S::Jacobi,cfs,plan::JacobiITransformPlan)
-#    @assert S==plan.space
-    jacobip(0:length(cfs)-1,S.a,S.b,tocanonical(S,plan.points))*cfs
-end
+*(P::JacobiITransformPlan,cfs) =
+    jacobip(0:length(cfs)-1,P.space.a,P.space.b,tocanonical(P.space,P.points))*cfs
+
 
 function coefficients(f::Vector,a::Jacobi,b::Chebyshev)
     if domain(a) == domain(b) && (!isapproxinteger(a.a-0.5) || !isapproxinteger(a.b-0.5))