+++ src/Spaces/Ultraspherical/Ultraspherical.jl
@@ -52,8 +52,8 @@ end
 
 function UltrasphericalPlan(λ::Number,vals)
     if λ == 0.5
-        cp=plan_transform(Chebyshev(),vals)
-        c2lp=FastTransforms.th_cheb2legplan(eltype(vals),length(vals))
+        cp = plan_transform(Chebyshev(),vals)
+        c2lp = FastTransforms.th_cheb2legplan(eltype(vals),length(vals))
         UltrasphericalPlan{typeof(cp),typeof(c2lp)}(cp,c2lp)
     else
         error("Not implemented")
@@ -71,18 +71,15 @@ function UltrasphericalIPlan(λ::Number,cfs)
 end
 
 *(UP::UltrasphericalPlan,v::AbstractVector) =
-    UP.cheb2legplan*transform(Chebyshev(),v,UP.chebplan)
+    UP.cheb2legplan*(UP.chebplan*v)
 *(UP::UltrasphericalIPlan,v::AbstractVector) =
-    itransform(Chebyshev(),UP.leg2chebplan*v,UP.chebiplan)
+    UP.chebiplan*(UP.leg2chebplan*v)
 
 
-plan_transform(S::Ultraspherical{Int},vals::Vector) = plan_transform(canonicalspace(S),vals)
+plan_transform(sp::Ultraspherical{Int},vals::Vector) = CanonicalTransformPlan(sp,vals)
 plan_transform(S::Ultraspherical,vals::Vector) = UltrasphericalPlan(order(S),vals)
-plan_itransform(S::Ultraspherical{Int},cfs::Vector) = plan_itransform(canonicalspace(S),cfs)
+plan_itransform(S::Ultraspherical{Int},cfs::Vector) = ICanonicalTransformPlan(sp,vals)
 plan_itransform(S::Ultraspherical,cfs::Vector) = UltrasphericalIPlan(order(S),cfs)
-
-transform(S::Ultraspherical,vals,pl::UltrasphericalPlan) = pl*vals
-itransform(S::Ultraspherical,cfs,pl::UltrasphericalIPlan) = pl*cfs
 
 ## Construction